Ensimmäinen valmis versio

+++ optical_tachometer.X/main.c
@@ -26,7 +26,7 @@ void SEGMENT_init(void);
 
 // Global variable to store adc result
 volatile uint16_t adcValue;
-// Global variable to store the rpm calculated from adcValue
+// Global variable to store the rpm calculated from rotations
 uint16_t rpm;
 // Variable to count rotations
 volatile uint16_t rotations;
@@ -35,7 +35,7 @@ volatile uint16_t rotations;
 volatile int msd;
 //int to check if the propeller is in front of the LDR, and makes sure, the 
 //propeller is only counted once
-uint8_t isPropOn;
+int isPropOn;
 
 // Fuction for sending text to computer terminal/putty
 static void USART0_sendChar(char c)
@@ -116,7 +116,7 @@ void RTC_init(void)
     // Enable periodic interrupt
     RTC.PITINTCTRL = RTC_PI_bm;
     // Selecting number of RTC clock cycles (16384) because we want interrupt
-    // twice a second?? and enable periodic interrupt timer
+    // twice a second and enable periodic interrupt timer
     RTC.PITCTRLA = RTC_PERIOD_CYC16384_gc | RTC_PITEN_bm;
 }
 void ADC_init(void)
@@ -149,9 +149,9 @@ void ADC_init(void)
 }
 void SEGMENT_init(void)
 {
-    // Tähä 7-segmenttihötskän tunnistus
+    // 7-segment display configurations
     
-    // portti c configured as an output
+    // Port C configured as an output
     VPORTC.DIR |= PIN0_bm;
     VPORTC.DIR |= PIN1_bm;
     VPORTC.DIR |= PIN2_bm;
@@ -167,16 +167,21 @@ ISR(RTC_PIT_vect)
 {
     // Clearing interrupt flag
     RTC.PITINTFLAGS = RTC_PI_bm;
-    // adcValue -> rpm
+    // rpm calculated from rotations (60 because reading value twice a second
+    // *120 but propeller has two wings /2 so *60)
     rpm = rotations*60;
+    // testing
     printf("%i rpm\r\n", rpm);
-    // rpm -> msd
+    // Counting first digit from rpm
     while(rpm >=10)
     {
+        // Dividing with 10 until only one number is left -> msd
         rpm = rpm / 10;
     }
     msd = rpm;
+    // Updating display to msd
     update_display(msd);
+    // Resetting rpm and rotations for next round
     rpm = 0;
     rotations=0;
 }
@@ -192,13 +197,13 @@ ISR(ADC0_RESRDY_vect)
 
 int main(void) 
 {
-    // Aluksi adc arvo vain 0
+    // In the beginning the value of adc is 0
     adcValue = 0;
-    // Aluksi rotations vain 0
+    // In the beginning the value of rotations is 0
     rotations = 0;
-    // Aluksi rpm alkuperäinen vain 0
+    // In the beginning the value of rpm is 0
     rpm = 0;
-    // Aluksi msd on vain 0
+    // In the beginning the value of msd is 0
     msd = 0;
     //In the beginning the value of isPropOn is 0
     isPropOn = 0;
@@ -218,11 +223,13 @@ int main(void)
     // Enable global interrupts
     sei();
     
+    // testing
     printf("Hello, testing testing.\r\n");
     
     while(1)
     {
-        //AdcValue while propeller is in front of LDR
+        //AdcValue while propeller is in front of LDR is 700 originally (will 
+        // be replaced later)
         while(adcValue>700)
         {   
             //makes sure the rotations are only updated once per rotation
@@ -236,6 +243,7 @@ int main(void)
             isPropOn=0;
         }
     }
+    // testing
     test();
     return 0;
 }
